Eta.Classes.Foldable

Documentation

class Foldable t where

Minimal complete definition

foldMap | foldr

Methods

foldMap :: Monoid m => (a -> m) -> t a -> m

foldr :: (a -> b -> b) -> b -> t a -> b

foldRight :: Foldable f => (a -> b -> b) -> b -> f a -> b

The Foldable type class defines that a type can have a foldRight operation which can be used to reduce it by passing a binary operation and a neutral element. This reduces starting from the right.

The List type is an example of Foldable

>>> foldRight (+) 0 [1, 2, 3]
6

Mnemonic: Reduceable

foldLeft :: Foldable f => (b -> a -> b) -> b -> f a -> b

foldLeft starts from the left, using the lazy evaluation capabilities of Eta. Note that this will get stuck in an infite loop if you pass an infite list to it.

strictFoldRight :: Foldable f => (a -> b -> b) -> b -> f a -> b

A version of foldRight that evaluates the operations inline, hence strict, the opposite of lazy.

strictFoldLeft :: Foldable f => (b -> a -> b) -> b -> f a -> b

A version of foldLeft that evaluates the operations inline, hence strict, the opposite of lazy.

monadicFoldRight :: (Foldable f, Monad m) => (a -> b -> m b) -> b -> f a -> m b

A version of foldRight that applies functions that are flatmappable, in the context of a type that implements a monad, and returns the result produced by the reduction of the structure, wrapped in that type:

>>> :{
 addIntoMaybe :: Int -> Int -> Maybe Int
 addIntoMaybe a b = Just (a + b)
:}

>>> monadicFoldRight addIntoMaybe 0 [1,2,3]
Just 6

monadicFoldLeft :: (Foldable f, Monad m) => (b -> a -> m b) -> b -> f a -> m b

Left-biased version of monadicFoldRight

toList :: Foldable f => f a -> [a]

Converts a type that implements Foldable into a list

isEmpty :: Foldable f => f a -> Bool

Checks if a Foldable structure is empty.

length :: Foldable f => f a -> Int

Returns the size of a structure.

isElementOf :: (Eq a, Foldable f) => a -> f a -> Bool

Checks if the element is contained in the structure.

maximum :: (Ord a, Foldable f) => f a -> Maybe a

Largest element in a structure. Returns Nothing if the structure is empty.

maximumBy :: Foldable f => (a -> a -> Ordering) -> f a -> Maybe a

Given some comparison function, return the maximum of a structure. Returns Nothing if the structure is empty.

unsafeMaximum :: (Ord a, Foldable f) => f a -> a

Warning: Partial functions should be avoided

Largest element in a structure. Errors if the structure is empty

minimum :: (Ord a, Foldable f) => f a -> Maybe a

Smallest element in a structure Returns Nothing if the structure is empty.

minimumBy :: Foldable f => (a -> a -> Ordering) -> f a -> Maybe a

Given some comparison function, return the minimum of a structure. Returns Nothing if the structure is empty.

unsafeMinimum :: (Ord a, Foldable f) => f a -> a

Warning: Partial functions should be avoided

Largest element in a structure. Errors if the structure is empty

sum :: (Num a, Foldable f) => f a -> a

Sum of the numbers of a structure

product :: (Num a, Foldable f) => f a -> a

Product of the numbers of a structure

findBy :: Foldable f => (a -> Bool) -> f a -> Maybe a

Given some predicate, findBy will return the first element that matches the predicate or Nothing if there is no such element

any :: Foldable f => (a -> Bool) -> f a -> Bool

Determines if any element satisfies the predicate

>>> any (== 1) [1, 2, 3]
True
>>> any (== 5) [1, 2, 3]
False

all :: Foldable f => (a -> Bool) -> f a -> Bool

Determines if all elements satisfy the predicate

>>> all (== 1) [1, 2, 3]
False
>>> all (< 5) [1, 2, 3]
True

discardTraverse :: (Foldable f, Applicative m) => (a -> m b) -> f a -> m ()

Sometimes we need to apply an action to each one of the elements. The function discardTraverse maps an action over each of the elements of the structure

>>> discardTraverse printLine ["Hello", "world", "!"]
Hello
world
!

foreach :: (Foldable f, Applicative m) => f a -> (a -> m b) -> m ()

Another alternative to discardTraverse is to use the foreach function, which is very familiar to a lot of developers.

>>> foreach [1..3] printShow
1
2
3