Eta.Classes.Monad

Description

The Monad type class defines that a type can have an flatMap operation which can be used to sequence actions over the same type.

All types that implement Monad must implement Applicative.

The List, Maybe, and IO types are examples of instances of Monad.

>>> flatMap (\x -> [1 .. x]) [1, 2, 3]
[1,1,2,1,2,3]

Mnemonic: Sequenceable

Documentation

class Applicative m => Monad m where

Minimal complete definition

(>>=)

Methods

(>>=) :: m a -> (a -> m b) -> m b

flatMap :: Monad m => (a -> m b) -> m a -> m b

(|>>) :: Monad m => m a -> (a -> m b) -> m b

Flatmaps the function of the right over the Monad of the left

>> readFile "data.txt" |>> printLine
"This is some data stored in data.txt"

(<<|) :: Monad m => (a -> m b) -> m a -> m b

Flatmaps the function of the left over the Monad of the left

>> printLine <<| readFile "data.txt"
"This is some data stored in data.txt"

Creates a new function that can be flatmapped by composing two functions:

 >> readAndPrint = readFile >=> printLine
 >> readAndPrint "data.txt"
 "This is some data stored in data.txt"
 

Sometimes referred as Kleisli composition

(>=>) :: Monad m => (a -> m b) -> (b -> m c) -> a -> m c

Inverted '(>=>)'

(<=<) :: Monad m => (b -> m c) -> (a -> m b) -> a -> m c