Eta.Classes.Monoid

Description

The Monoid type class defines that a type can have an append operation which can be used to append two things of the same type

The List, and String types are examples of instances of Monoid.

Mnemonic: Appendable

Documentation

class Monoid a where

Minimal complete definition

mempty, mappend

Methods

mempty :: a

mappend :: a -> a -> a

mconcat :: [a] -> a

append :: Monoid a => a -> a -> a

Appends two values.

>>> append "Hello " "world!"
"Hello world!"

neutral :: Monoid a => a

A Monoid also requires that the type has a neutral element that does not affect concatenation.

$ >>> import Prelude (String, Int)

>>> neutral :: String
""

>>> neutral :: [Int]
[]

concat :: Monoid a => [a] -> a

You can reduce a list of elements that implement the Monoid type class by using concat

>>> import Prelude ()
>>> concat ["Hello ", "world", "!"]
"Hello world!"

(<+>) :: Monoid a => a -> a -> a

Operator for the append operation

>>> "Hello " <+> "world!"
"Hello world!"