Eta.Classes.Read

Description

The Read type class defines an interface with a method read that allows converting a String into a value. Think of this as a naive form of deserialization.

Documentation

>>> import Prelude(Int)

read :: Read a => String -> Maybe a

Tries to deserialize a value using the Read type class. If it fails, returns Nothing

>>> read "1" :: Maybe Int
Just 1
>>> read "pizza" :: Maybe Int
Nothing

unsafeRead :: Read a => String -> a

Tries to deserialize a value using the Read type class. If it fails, it throws an error

>>> unsafeRead "1" :: Int
1

class Read a where

Minimal complete definition

readsPrec | readPrec

Methods

readsPrec :: Int -> ReadS a