Eta.Classes.Real

Description

The Real type class defines how to convert a numerical value into a full precision Rational value

Documentation

class (Num a, Ord a) => Real a where

Minimal complete definition

toRational

Methods

toRational :: a -> Rational