Eta.Classes.Show

Description

The Show type class defines an interface with a method show that allows converting a value into a String

Documentation

Converts a value into a String

>>> show 1
"1"

show :: Show a => a -> String