Eta.Prelude.Classes.Enum

Description

The Enum class defines operations on sequentially ordered types

Documentation

successor :: Enum a => a -> a

Successor of a value

>>> successor 1
2

predecessor :: Enum a => a -> a

Predecessor of a value

>>> predecessor 1
0

toEnumeration :: Enum a => Int -> a

Converts an Int into the equivalent value of the enumeration

>>> toEnumeration 1 :: Bool
True

fromEnumeration :: Enum a => a -> Int

Converts a value from an enumeration into the equivalent Int value

>>> fromEnumeration False
0

enumerateFrom :: Enum a => a -> [a]

Generates a list from the beginning of the type to the end

>>> take 4 (enumerateFrom 0)
[0,1,2,3]
>>> enumerateFrom False
[False,True]

steppingEnumerateFrom :: Enum a => a -> a -> [a]

Generates a list from the beginning of the type to the end, with steps

>>> take 4 (steppingEnumerateFrom 0 2)
[0,2,4,6]

enumerateFromTo :: Enum a => a -> a -> [a]

Generates a list from the beginning of the type to the specified limit

>>> enumerateFromTo 0 3
[0,1,2,3]

steppingEnumerateFromTo :: Enum a => a -> a -> a -> [a]

Generates a list from the beginning of the type to the specified limit, with steps

>>> steppingEnumerateFromTo 0 2 8
[0,2,4,6,8]