Eta.Prelude.Classes.Floating

Description

The Floating type class defines trigonometric and hyperbolic functions

Documentation

class Fractional a => Floating a where

Minimal complete definition

pi, exp, log, sin, cos, asin, acos, atan, sinh, cosh, asinh, acosh, atanh

Methods

pi :: a

exp :: a -> a

log :: a -> a

sqrt :: a -> a

(**) :: a -> a -> a

logBase :: a -> a -> a

sin :: a -> a

cos :: a -> a

tan :: a -> a

asin :: a -> a

acos :: a -> a

atan :: a -> a

sinh :: a -> a

cosh :: a -> a

tanh :: a -> a

asinh :: a -> a

acosh :: a -> a

atanh :: a -> a