Eta.Prelude.Classes.IsString

Description

The type class IsString defines a convertString method that allows to convert from a regular String to another type that implements this type class.

It is very useful for using with other string-like types like ByteString or ByteArray

Documentation

convertString :: IsString a => String -> a

Converts a String into another string-like type.