Eta.Prelude.Classes.Real

Description

The Real type class defines how to convert a numerical value into a full precision Rational value

Documentation

toRational :: Real a => a -> Rational

The rational equivalent of its argument