Eta.Prelude.Core

Operators

Flow and application

(|>) :: a -> (a -> b) -> b

Pipe operator. Functions can be applied using the pipe operator, reducing parentheses and increasing the code readability.

Instead of writing

  (printLine (join ", " (sort names)))

Use the pipe operator:

  names
    |> sort
    |> join ", "
    |> printLine

(<|) :: (a -> b) -> a -> b

Pipe backwards operator. Applies the function of the left to the value of the right.

>>> (+ 2) <| 1
3

($) :: (a -> b) -> a -> b

Application operator. Synonymous to (<|), useful for skipping a level of parentheses.

It can be read as "enclose in parentheses from here to the end of the expression."

>>> (print $ show 4)
4

(>>>) :: (a -> b) -> (b -> c) -> a -> c

Forward function composition.

Creates a new function that acts as a "pipe", passing the arguments to the first one and then to the second:

>>> addOne x = x + 1
>>> subTwo x = x - 2
>>> subOne = addOne >>> subTwo
>>> subOne 4
3

(<<<) :: (a -> b) -> (b -> c) -> a -> c

Backwards function composition.

Creates a new function that acts as a "pipe", passing the arguments to the second one and then to the first:

>>> addOne x = x + 1
>>> subTwo x = x - 2
>>> subOne = subTwo <<< addOne
>>> subOne 4
3

Boolean

(&&) :: Bool -> Bool -> Bool

Boolean and

>>> True && True
True
>>> True && False
False

(||) :: Bool -> Bool -> Bool

Boolean or

>>> True || True
True
>>> True || False
True

Functions

General functions

first :: (a, b) -> a

Extract the first component of a pair

>>> first (1, 2)
1

second :: (a, b) -> b

Extract the second component of a pair

>>> second (1, 2)
2

identity :: a -> a

The identity function, returns the passed argument as it is

>>> identity 1
1

getOrElse :: Maybe a -> a -> a

Takes a default value and a Maybe value. If the Maybe is Nothing, it returns the default value, returns the contents of the Maybe >>> let x = Just Something >>> x getOrElse "Nothing found!" Something

>>> let y = Nothing
>>> y `getOrElse` "Nothing found!"
"Nothing found!"

handleMaybe :: b -> (a -> b) -> Maybe a -> b

Handler for the Maybe type Takes a default value, a function and a Maybe value. If the Maybe is Nothing, it returns the default value, otherwise, it maps the function and returns the result

>>> let myMaybe = Just 42
>>> handleMaybe "Nothing found" (\x -> "Got: " <+> show x) myMaybe
"Got: 42"

>>> let myMaybe' = Nothing
>>> handleMaybe "Nothing found" (\x -> "Got: " <+> show x) myMaybe'
"Nothing found"

handleEither :: (a -> c) -> (b -> c) -> Either a b -> c

Handler for the Either type Takes two functions and an Either.

If the value is a 'Left a' apply the first function to a. If the value is a 'Right b' apply the first function to b.

flip :: (a -> b -> c) -> b -> a -> c

Flips the arguments of a function

>>> subtract a b = a - b
>>> inverseSubtract = flip subtract
>>> subtract 3 2
1
>>> inverseSubtract 3 2
-1

constantly :: a -> b -> a

Constructs a function that returns the same output for the same input

>>> universe = constantly 42
>>> universe 35
42
>>> map (constantly 10) [1..3]
[10,10,10]

die :: [Char] -> a

Warning: die detected: Partial functions should be avoided

Stops execution and displays an error message.

uncurry :: (a -> b -> c) -> (a, b) -> c

Converts a two argument function in a function that accepts a pair

>>> addElements = uncurry (+)
>>> addElements (1,2)
3

curry :: ((a, b) -> c) -> a -> b -> c

Converts a function that accepts a pair in a two argument function

>>> firstArgument = curry first
>>> firstArgument 1 2
1

Boolean

not :: Bool -> Bool

Boolean not

>>> not True
False

otherwise :: Bool

otherwise is defined as True.

Used for guards: >>> :{ factorial n | n == 0 = 1 | otherwise = n * factorial (n - 1) :}

>>> factorial 3
6

IO operations

print :: String -> IO ()

Print a String on screen

printLine :: String -> IO ()

Print a String on screen, followed by a newline.

printShow :: String -> IO ()

Converts the value to a String using show and then printLines it.

getArgs :: IO [String]

Returns the arguments of the program