Eta.Classes.Functor

Description

The Functor type class defines that a type can have a map operation which can be used to map a function over the elements of it.

The list and the Maybe types are examples of instances of Functor.

Mnemonic: Mappable

Documentation

map :: Functor f => (a -> b) -> f a -> f b

Maps a function over a value

>>> map (+1) [1,2,3]
[2,3,4]

Map forward

(|$>) :: Functor f => f a -> (a -> b) -> f b

Maps the function of the right to the Functor of the left:

>>> [1,2,3] |$> (+ 1)
[2,3,4]

Map backwards

(<$|) :: Functor f => (a -> b) -> f a -> f b

Maps the function of the left to the Functor of the right:

>>> (+1) <$| [1,2,3]
[2,3,4]

Discard

discard :: Functor f => f a -> f ()

Discards the value inside of the functor.

>>> discard [1,2,3]
[(),(),()]