Eta.Classes.Show

Description

The Show type class defines an interface with a method show that allows converting a value into a String

Documentation

class Show a where

Minimal complete definition

showsPrec | show

Methods

showsPrec :: Int -> a -> ShowS

show :: a -> String

showList :: [a] -> ShowS