Eta.Core

Description

STUFF HERE IS WORK IN PROGRESS, IGNORE THIS MODULE

Operators

Flow and application

(|>) :: a -> (a -> b) -> b

Pipe operator. Functions can be applied using the pipe operator, reducing parentheses and increasing the code readability.

Instead of writing

  (printLine (join ", " (sort names)))

Use the pipe operator:

  names
    |> sort
    |> join ", "
    |> printLine

(<|) :: (a -> b) -> a -> b

Pipe backwards operator. Applies the function of the left to the value of the right.

>>> (+ 2) <| 1
3

($) :: (a -> b) -> a -> b

Application operator. Synonymous to (<|), useful for skipping a level of parentheses.

It can be read as "enclose in parentheses from here to the end of the expression."

>>> (print $ show 4)
4

(>>>) :: (a -> b) -> (b -> c) -> a -> c

Forward function composition.

Creates a new function that acts as a "pipe", passing the arguments to the first one and then to the second:

>>> addOne x = x + 1
>>> subTwo x = x - 2
>>> subOne = addOne >>> subTwo
>>> subOne 4
3

(<<<) :: (a -> b) -> (b -> c) -> a -> c

Backwards function composition.

Creates a new function that acts as a "pipe", passing the arguments to the second one and then to the first:

>>> addOne x = x + 1
>>> subTwo x = x - 2
>>> subOne = subTwo <<< addOne
>>> subOne 4
3

Boolean

(&&) :: Bool -> Bool -> Bool

Boolean and

>>> True && True
True
>>> True && False
False

(||) :: Bool -> Bool -> Bool

Boolean or

>>> True || True
True
>>> True || False
True

Functions

General functions

first :: (a, b) -> a

Extract the first component of a pair

>>> first (1, 2)
1

second :: (a, b) -> b

Extract the second component of a pair

>>> second (1, 2)
2

identity :: a -> a

The identity function, returns the passed argument as it is

>>> identity 1
1

flip :: (a -> b -> c) -> b -> a -> c

Flips the arguments of a function

>>> subtract a b = a - b
>>> inverseSubtract = flip subtract
>>> subtract 3 2
1
>>> inverseSubtract 3 2
-1

constantly :: a -> b -> a

Constructs a function that returns the same output for the same input

>>> universe = constantly 42
>>> universe 35
42
>>> map (constantly 10) [1..3]
[10,10,10]

die :: [Char] -> a

Warning: die detected: Partial functions should be avoided

Stops execution and displays an error message.

uncurry :: (a -> b -> c) -> (a, b) -> c

Converts a two argument function in a function that accepts a pair

>>> addElements = uncurry (+)
>>> addElements (1,2)
3

curry :: ((a, b) -> c) -> a -> b -> c

Converts a function that accepts a pair in a two argument function

>>> firstArgument = curry first
>>> firstArgument 1 2
1

swap :: (a, b) -> (b, a)

Swaps the components of a pair

>>> swap (1,2)
(2,1)

undefined :: a

A value that can be of any type. The compiler will recognize this and insert an error message appropriate to the context where it appears.

Usually, also called Bottom

stopIfUndefined :: a -> b -> b

Returns undefined if a is undefined, otherwise returns b

>>> stopIfUndefined 1 2
2

Boolean

not :: Bool -> Bool

Boolean not

>>> not True
False

IO operations

print :: String -> IO ()

Print a String on screen

printLine :: String -> IO ()

Print a String on screen, followed by a newline.

printShow :: String -> IO ()

Converts the value to a String using show and then printLines it.

getArgs :: IO [String]

Returns the arguments of the program