Eta.Types.Maybe

Description

The Maybe type encapsulates an optional value.

It either contains a value, represented by Just a, or not, represented by Nothing.

Documentation

data Maybe a :: * -> *

Constructors

Nothing

Just a

isJust :: Maybe a -> Bool

Returns True if the Maybe is a Just

isNothing :: Maybe a -> Bool

Returns True if the Maybe is a Nothing

unsafeGetValue :: Maybe a -> a

Warning: Partial functions should be avoided.

Extracts the element out of a Just. Throws an error if it is Nothing

>>> unsafeGetValue (Just 1)
1

getOrElse :: Maybe a -> a -> a

Takes a default value and a Maybe value. If the Maybe is Nothing, it returns the default value, returns the contents of the Maybe

>>> x = Just "Something"
>>> x `getOrElse` "Nothing found!"
"Something"

>>> y = Nothing
>>> y `getOrElse` "Nothing found!"
"Nothing found!"

handleMaybe :: b -> (a -> b) -> Maybe a -> b

Handler for the Maybe type Takes a default value, a function and a Maybe value. If the Maybe is Nothing, it returns the default value, otherwise, it maps the function and returns the result

>>> myMaybe = Just 42
>>> handleMaybe "Nothing found" (\x -> "Got: " <+> show x) myMaybe
"Got: 42"

>>> myMaybe' = Nothing
>>> handleMaybe "Nothing found" (\x -> "Got: " <+> show x) myMaybe'
"Nothing found"